ExplainableChess-main/requirements.txt
python-chess
stockfish
numpy
pandas
tqdm
torch
transformers
datasets
accelerate
huggingface-hub
